Refactoring format date and time
